+++ lucene/src/test/org/apache/lucene/search/payloads/TestPayloadTermQuery.java
@@ -336,7 +336,7 @@ public class TestPayloadTermQuery extends LuceneTestCase {
         }
 
         @Override
-        public float idf(int docFreq, int numDocs) {
+        public float idf(long docFreq, long numDocs) {
           return 1;
         }
